# admin-ui/openapi.yaml
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "JETTY",
                 "version":  "2.0.0"
             },
    "paths":  {
...
                  "/api/chat":  {
                                    "post":  {
                                                 "summary":  "Chat",
                                                 "operationId":  "chat_api_chat_post",
                                                 "requestBody":  {
                                                                     "content":  {
                                                                                     "application/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/ChatRequest"
                                                                                                                         }
                                                                                                          }
                                                                                 },
                                                                     "required":  true
                                                                 },
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Successful Response",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {

                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                                   "422":  {
                                                                               "description":  "Validation Error",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           }
                                                               }
                                             }
                                },
                  "/api/remember":  {
                                        "post":  {
                                                     "summary":  "Remember",
                                                     "operationId":  "remember_api_remember_post",
                                                     "requestBody":  {
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/RememberRequest"
                                                                                                                             }
                                                                                                              }
                                                                                     },
                                                                         "required":  true
                                                                     },
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "Successful Response",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {

                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               },
                                                                       "422":  {
                                                                                   "description":  "Validation Error",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               }
                                                                   }
                                                 }
                                    },
...
              },
    "components":  {
                       "schemas":  {
                                       "ChatRequest":  {
                                                           "properties":  {
                                                                              "message":  {
                                                                                              "type":  "string",
                                                                                              "maxLength":  4000,
                                                                                              "minLength":  1,
                                                                                              "title":  "Message"
                                                                                          },
                                                                              "session_id":  {
                                                                                                 "type":  "string",
                                                                                                 "maxLength":  96,
                                                                                                 "title":  "Session Id",
                                                                                                 "default":  "default"
                                                                                             },
                                                                              "voice_id":  {
                                                                                               "anyOf":  [
                                                                                                             {
                                                                                                                 "type":  "string",
                                                                                                                 "maxLength":  120
                                                                                                             },
                                                                                                             {
                                                                                                                 "type":  "null"
                                                                                                             }
                                                                                                         ],
                                                                                               "title":  "Voice Id"
                                                                                           },
                                                                              "model_provider":  {
                                                                                                     "anyOf":  [
                                                                                                                   {
                                                                                                                       "type":  "string",
                                                                                                                       "maxLength":  32
                                                                                                                   },
                                                                                                                   {
                                                                                                                       "type":  "null"
                                                                                                                   }
                                                                                                               ],
                                                                                                     "title":  "Model Provider"
                                                                                                 }
                                                                          },
                                                           "type":  "object",
                                                           "required":  [
                                                                            "message"
                                                                        ],
                                                           "title":  "ChatRequest"
                                                       },
...
                                   }
                   }
}